add custom pid, fix tests

+++ host/math/src/common.rs
@@ -198,7 +198,7 @@ pub fn compute_ntf_thermistor_temperature(
     r0: Resistance,
     r_series: Resistance,
 ) -> Temperature {
-    let r_ntc = r_series * (sample / (max_sample - sample)) as f64;
+    let r_ntc = r_series * sample as f64 / (max_sample - sample) as f64;
     let val_inv =
         (1.0 / t0.as_kelvin()) + (1.0 / b.as_kelvin()) * (((r_ntc / r0) as f32).ln() as f64);
     Temperature::from_kelvin(1.0 / val_inv)
